backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
python-engineio==4.8.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
redis==5.0.1
pyaudio==0.2.14
numpy==1.26.0
scipy==1.11.0
websockets==12.0
aiofiles==23.2.1
jinja2==3.1.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.13.0
sqlalchemy==2.0.0
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.25.0
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
